fix(striker-ui-api): migrate job to access interact
