+++ striker-ui-api/src/lib/request_handlers/file/buildQueryFileDetail.ts
@@ -1,7 +1,9 @@
+import { DELETED } from '../../consts';
+
 import join from '../../join';
 import { stdoutVar } from '../../shell';
 
-const buildQueryFileDetail = ({
+export const buildQueryFileDetail = ({
   fileUUIDs = ['*'],
 }: {
   fileUUIDs?: string[] | '*';
@@ -34,8 +36,6 @@ const buildQueryFileDetail = ({
       ON fil.file_uuid = fil_loc.file_location_file_uuid
     JOIN anvils AS anv
       ON fil_loc.file_location_anvil_uuid = anv.anvil_uuid
-    WHERE fil.file_type != 'DELETED'
+    WHERE fil.file_type != '${DELETED}'
       ${condFileUUIDs};`;
 };
-
-export default buildQueryFileDetail;
