+++ striker-ui-api/src/lib/request_handlers/file/getFileDetail.ts
@@ -1,12 +1,10 @@
 import { RequestHandler } from 'express';
 
 import buildGetRequestHandler from '../buildGetRequestHandler';
-import buildQueryFileDetail from './buildQueryFileDetail';
+import { buildQueryFileDetail } from './buildQueryFileDetail';
 import { sanitizeSQLParam } from '../../sanitizeSQLParam';
 
-const getFileDetail: RequestHandler = buildGetRequestHandler(
+export const getFileDetail: RequestHandler = buildGetRequestHandler(
   ({ params: { fileUUID } }) =>
     buildQueryFileDetail({ fileUUIDs: [sanitizeSQLParam(fileUUID)] }),
 );
-
-export default getFileDetail;
