+++ striker-ui-api/src/lib/request_handlers/file/index.ts
@@ -1,0 +1,5 @@
+export * from './createFile';
+export * from './deleteFile';
+export * from './getFile';
+export * from './getFileDetail';
+export * from './updateFile';
