+++ striker-ui-api/src/lib/request_handlers/host/deleteHostConnection.ts
@@ -10,33 +10,31 @@ export const deleteHostConnection: RequestHandler<
   unknown,
   undefined,
   DeleteHostConnectionRequestBody
-> = (request, response) => {
+> = async (request, response) => {
   const { body } = request;
-  const hostUUIDs = Object.keys(body);
+  const hostUuids = Object.keys(body);
 
-  hostUUIDs.forEach((key) => {
-    const hostUUID = toHostUUID(key);
-    const peerHostUUIDs = body[key];
+  for (const key of hostUuids) {
+    const hostUuid = toHostUUID(key);
+    const peerHostUuids = body[key];
 
-    peerHostUUIDs.forEach((peerHostUUID) => {
+    for (const peerHostUuid of peerHostUuids) {
       try {
-        job({
+        await job({
           file: __filename,
-          job_command: `${SERVER_PATHS.usr.sbin['striker-manage-peers'].self} --remove --host-uuid ${peerHostUUID}`,
+          job_command: `${SERVER_PATHS.usr.sbin['striker-manage-peers'].self} --remove --host-uuid ${peerHostUuid}`,
           job_description: 'job_0014',
-          job_host_uuid: hostUUID,
+          job_host_uuid: hostUuid,
           job_name: 'striker-peer::delete',
           job_title: 'job_0013',
         });
       } catch (subError) {
-        stderr(`Failed to delete peer ${peerHostUUID}; CAUSE: ${subError}`);
+        stderr(`Failed to delete peer ${peerHostUuid}; CAUSE: ${subError}`);
 
-        response.status(500).send();
-
-        return;
+        return response.status(500).send();
       }
-    });
-  });
+    }
+  }
 
   response.status(204).send();
 };
