+++ striker-ui-api/src/lib/request_handlers/host/setHostInstallTarget.ts
@@ -4,36 +4,38 @@ import { LOCAL } from '../../consts/LOCAL';
 import SERVER_PATHS from '../../consts/SERVER_PATHS';
 
 import { job } from '../../accessModule';
-import { stderr, stdout } from '../../shell';
+import { stderr, stdoutVar } from '../../shell';
 
-export const setHostInstallTarget: RequestHandler = (request, response) => {
-  stdout(
-    `Begin set host install target.\n${JSON.stringify(request.body, null, 2)}`,
-  );
+export const setHostInstallTarget: RequestHandler<
+  UpdateHostParams,
+  undefined,
+  SetHostInstallTargetRequestBody
+> = async (request, response) => {
+  const { body, params } = request;
 
-  const { isEnableInstallTarget } =
-    request.body as SetHostInstallTargetRequestBody;
-  const { hostUUID: rawHostUUID } = request.params as UpdateHostParams;
-  const hostUUID: string | undefined =
-    rawHostUUID === LOCAL ? undefined : rawHostUUID;
+  stdoutVar(body, `Begin set host install target; body=`);
+
+  const { isEnableInstallTarget } = body;
+  const { hostUUID: rHostUuid } = params;
+
+  const hostUuid: string | undefined =
+    rHostUuid === LOCAL ? undefined : rHostUuid;
   const task = isEnableInstallTarget ? 'enable' : 'disable';
 
   try {
-    job({
+    await job({
       file: __filename,
       job_command: `${SERVER_PATHS.usr.sbin['striker-manage-install-target'].self} --${task}`,
       job_description: 'job_0016',
-      job_host_uuid: hostUUID,
+      job_host_uuid: hostUuid,
       job_name: `install-target::${task}`,
       job_title: 'job_0015',
     });
   } catch (subError) {
     stderr(`Failed to ${task} install target; CAUSE: ${subError}`);
 
-    response.status(500).send();
-
-    return;
+    return response.status(500).send();
   }
 
-  response.status(200).send();
+  response.status(204).send();
 };
