+++ striker-ui-api/src/types/DBInsertOrUpdateFunctionCommon.d.ts
@@ -2,8 +2,3 @@ type DBInsertOrUpdateFunctionCommonParams = ModuleSubroutineCommonParams & {
   file: string;
   line?: number;
 };
-
-type DBInsertOrUpdateFunctionCommonOptions = Omit<
-  ExecModuleSubroutineOptions,
-  'subParams' | 'subModuleName'
->;
