+++ striker-ui-api/src/types/DBInsertOrUpdateJobFunction.d.ts
@@ -7,5 +7,3 @@ type DBJobParams = DBInsertOrUpdateFunctionCommonParams & {
   job_host_uuid?: string;
   job_progress?: number;
 };
-
-type DBInsertOrUpdateJobOptions = DBInsertOrUpdateFunctionCommonOptions;
